+++ src/main/java/com/seong/shoutlink/domain/hub/repository/HubMemberRepositoryImpl.java
@@ -1,7 +1,7 @@
-package com.seong.shoutlink.domain.hubmember.repository;
+package com.seong.shoutlink.domain.hub.repository;
 
 import com.seong.shoutlink.domain.hub.Hub;
-import com.seong.shoutlink.domain.hubmember.service.HubMemberRepository;
+import com.seong.shoutlink.domain.hub.service.HubMemberRepository;
 import com.seong.shoutlink.domain.member.Member;
 import org.springframework.stereotype.Repository;
