+++ src/main/java/com/seong/shoutlink/domain/hub/repository/HubMemberRole.java
@@ -1,4 +1,4 @@
-package com.seong.shoutlink.domain.hubmember;
+package com.seong.shoutlink.domain.hub.repository;
 
 public enum HubMemberRole {
     MASTER, PARTICIPANTS
