+++ src/main/java/com/seong/shoutlink/domain/hub/service/HubMemberRepository.java
@@ -1,4 +1,4 @@
-package com.seong.shoutlink.domain.hubmember.service;
+package com.seong.shoutlink.domain.hub.service;
 
 import com.seong.shoutlink.domain.hub.Hub;
 import com.seong.shoutlink.domain.member.Member;
