+++ src/main/java/com/seong/shoutlink/domain/linkbundle/service/LinkBundleService.java
@@ -4,7 +4,7 @@ import com.seong.shoutlink.domain.exception.ErrorCode;
 import com.seong.shoutlink.domain.exception.ShoutLinkException;
 import com.seong.shoutlink.domain.hub.Hub;
 import com.seong.shoutlink.domain.hub.service.HubRepository;
-import com.seong.shoutlink.domain.hubmember.service.HubMemberRepository;
+import com.seong.shoutlink.domain.hub.service.HubMemberRepository;
 import com.seong.shoutlink.domain.linkbundle.HubLinkBundle;
 import com.seong.shoutlink.domain.linkbundle.LinkBundle;
 import com.seong.shoutlink.domain.linkbundle.MemberLinkBundle;
