+++ src/test/java/com/seong/shoutlink/domain/hubmember/repository/StubHubMemberRepository.java
@@ -1,7 +1,7 @@
 package com.seong.shoutlink.domain.hubmember.repository;
 
 import com.seong.shoutlink.domain.hub.Hub;
-import com.seong.shoutlink.domain.hubmember.service.HubMemberRepository;
+import com.seong.shoutlink.domain.hub.service.HubMemberRepository;
 import com.seong.shoutlink.domain.member.Member;
 import java.util.Arrays;
 import java.util.HashMap;
